+++ src/components/EyebrowLabel/index.ts
@@ -1,0 +1,2 @@
+export { EyebrowLabel } from './EyebrowLabel'
+export type { EyebrowLabelProps } from './EyebrowLabel'