+++ src/components/Modal/Modal.tsx
@@ -156,7 +156,7 @@ const ModalDescription = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <DialogPrimitive.Description
     ref={ref}
-    className={cn("text-sm text-gray-600 dark:text-navy-400", className)}
+    className={cn("text-sm text-gray-600 dark:text-navy-300", className)}
     {...props}
   />
 ));